make requested refactors
