+++ include/scea/algo/bfs.hpp
@@ -1,23 +1,14 @@
 #pragma once
 
+#include "algo_interface.hpp"
+
 #include <limits>
 #include <atomic>
 #include <memory>
 
-#include "graph.hpp"
 #include "galois/LargeArray.h"
 
-namespace scea {
-
-class Algo {
-public:
-  virtual void operator()(Graph& g) = 0;
-};
-
-class Nop : public Algo {
-public:
-  void operator()(Graph&) override {}
-};
+namespace scea::algo {
 
 class SSSP_BFS : public Algo {
   const uint64_t m_src;
@@ -25,8 +16,8 @@ class SSSP_BFS : public Algo {
 public:
   SSSP_BFS(uint64_t src) : m_src(src) {}
 
-  static std::unique_ptr<galois::LargeArray<uint64_t>> compute(Graph& g,
-                                                               uint64_t src) {
+  static galois::LargeArray<uint64_t> compute(scea::graph::MutableGraph& g,
+                                              uint64_t src) {
     using Cont = galois::InsertBag<uint64_t>;
 
     auto currSt = std::make_unique<Cont>();
@@ -35,19 +26,17 @@ public:
     galois::LargeArray<std::atomic_bool> visited;
     visited.create(g.size(), false);
 
-    auto shortest_path = std::make_unique<galois::LargeArray<uint64_t>>();
-    shortest_path->create(g.size(), std::numeric_limits<uint64_t>::max());
+    galois::LargeArray<uint64_t> shortest_path;
+    shortest_path.create(g.size(), std::numeric_limits<uint64_t>::max());
 
     nextSt->push(src);
-    visited[src]           = true;
-    shortest_path->at(src) = 0U;
+    shortest_path[src] = 0U;
 
     uint64_t level = 0U;
 
     while (!nextSt->empty()) {
       std::swap(currSt, nextSt);
       nextSt->clear();
-      ++level;
 
       galois::do_all(
           galois::iterate(*currSt),
@@ -55,7 +44,7 @@ public:
             if (visited[vertex].exchange(true, std::memory_order_seq_cst))
               return;
 
-            shortest_path->at(vertex) = level;
+            shortest_path[vertex] = level;
 
             // not previously visited, add all edges
             g.for_each_edge(vertex, [&](uint64_t const& neighbor) {
@@ -64,11 +53,13 @@ public:
             });
           },
           galois::steal());
+
+      ++level;
     }
     return shortest_path;
   }
 
-  void operator()(Graph& g) override { compute(g, m_src); }
+  void operator()(scea::graph::MutableGraph& g) override { compute(g, m_src); }
 };
 
-} // namespace scea
+} // namespace scea::algo
